feat(auth): abandon frontend-only authentication solution
The "angular-oauth2-oidc" library automatically generates code_challenge
and code_verifier, passing them to the authentication server. However,
obtaining a token requires client_secret, which cannot be stored on the
frontend. This forces part of the process to be moved to the backend.
Since the library doesn't allow passing code_verifier to the backend,
this approach was abandoned.